Solved bug where stopping listening would shut down the servers
The reasont that this bug was occuring was that the processes that were
managing the gunicorn server were child processes of the python server.
As a result, when a SIGINT message was sent (regardless of whether or
not it was inside the listen mode or not, even after removing the
handler temporarily. It would shut down the wsgi servers. As a result,
the wsgi servers are now separate processes.. Further to also solve
another bug that was making this bug harder to diagnose. I made it such
that when any domain is modified , so either created, paused, or removed
I save to the domain.json (which I really should have done in the first
place)