Solved server timing out issue, and ports still being occupied by leftover processes
I was super silly here. Foremost, the server had limited workers, and
the chunks that were being sent by the agent were overwhelming it (poor
guy). Increased the amount of workers for wsgi and decreased the TTL,
also made it such that the agent sends its message in a single post
message. If it is using chunks, it leverages the same post session. That
seemed to fix it for now. Further, there were still leftover processes,
I finally realized (duh) that multiple processes are spawned. So keeping
track of the PID makes no sense, and removing by the PID makes no
senese. Instread added a new fucntion to the wsgi_creator that removes
all processes on a given port. okay we should be good now, hopefully I
can move onto more interesting stuff