Added better testing functionality for the domain_orchestrator, wsgi_creator, and agent modules. This further involved doing some dependency injection stuff for files properly, particularly for the wsgi_creator and its files.
I realized that I was struggling to test the agent.py properly as I had
not properly done good dependency injection for the wsgi_creator class.
I made it such that it now takes the directories where it stores its own
files (as well as the directory that it injects into the flask
application that it runs in wsgi files on startup) as well as a
redis_client that is able to connect to a fake server if needed for
testing.