Added testing functionality for the agent to recieve commands, specifically this required some refactory of the wsgi_creator to actually take in the dependencies that needed to be passed ito the flask application that is instantiated in its generated file
There were a few issues here. The largest being that during testing, the
redis clients were matching to different servers. This is because one of
them was a fake redis client, and then the one actually being
instantiated inside of the flask application was a real edis client
solution here was pretty straightforward, using a fake TCP server that
is provided, then pass through into the flask application that exists in
the corresponding wsgi file. I also found another error that literally
took 4 hours to understand. simply, it is that fake redis for some
reaosn cannot handle new line characters being passed into it's streams,
which was interesting. Reporting the bug on github