Added flask endpoint and agent request and parsing functionality to apply modification commands
Still need to apply the sending of the modification commands back, as
well as the TUI functinoality inside of main.py in order to handle this.
Further I still need to run an entire end to end integration test using
this endpoint. It is tested in the flask application however