Added first integration test between the go agent and the WSGI python servers
Encountered a bug where the nginx conf files would not copy over. This
is becuase for some reason the /servers directory inside of nginx was
deleted. The issue was solved when I added back the server. I made it
inside of hte nginx_controller such tht if the /servers directory does
not exist it creates it. In the future, to make this cross comptabile
across different systems. It may need to be such that the .conf file in
nginx is set to recieve. Maybe this requires being a group of
orchestrated containers. Who knows.